AppBarを切り分けて共通化し、それぞれのページで好きなAppBarを使えるようにした。
