+++ lib/tag_library/library_page.dart
@@ -119,8 +119,9 @@ class TagCard extends StatelessWidget {
 
           // if (false) const CircularProgressIndicator(),
           ListTile(
-            visualDensity: VisualDensity(horizontal: 0, vertical: -4),
-            contentPadding: EdgeInsets.symmetric(vertical: 0, horizontal: 0),
+            visualDensity: const VisualDensity(horizontal: 0, vertical: -4),
+            contentPadding:
+                const EdgeInsets.symmetric(vertical: 0, horizontal: 0),
             title: Text(tag.tagName),
             subtitle: Text(
               '${tag.photoIdList.length}個',
